services/user_service/pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aio-pika>=9.5.7",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.119.1",
    "pydantic-settings>=2.11.0",
    "pyjwt[crypto]>=2.10.1",
    "sqlalchemy>=2.0.44",
]
